% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_power.R
\name{simulate.wolfData}
\alias{simulate.wolfData}
\title{simulate.wolfData}
\usage{
\method{simulate}{wolfData}(sample.size, effMF, effDens)
}
\arguments{
\item{sample.size}{sample size}

\item{effMF}{sex effect size}

\item{effDens}{wolf density effect size}
}
\value{

}
\description{
Simulate data with random variance for wolves and locations
}
\details{
Simulate data...
}
\author{
Emanuel Heitlinger
}
